Add version-specific handling that forward cdf.catalogTableOpt for Delta 4.0/4.1
